pyproject.toml
[tool.poetry]
name = "privotron"
version = "1.0"
description = "Crowd-sourced opt-out templates for data brokers, helping protect your privacy."
authors = ["kloeffler95 <kevin.loeffler@colorado.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
click = "^8.1.8"
playwright = "^1.52.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
